+++ build-logic/build-logic-base/src/main/kotlin/com/ensody/buildlogic/KotlinMultiplatformExt.kt
@@ -83,7 +83,6 @@ fun KotlinMultiplatformExtension.applyKmpHierarchy(block: KotlinHierarchyBuilder
                 withWasmJs()
                 withWasmWasi()
                 withIos()
-                withMacos()
                 withJvm()
                 withAndroidTarget()
             }