+++ build-logic/build-logic/src/main/kotlin/com/ensody/buildlogic/BuildLogicPlugin.kt
@@ -102,6 +102,7 @@ fun Project.setupBuildLogic(block: Project.() -> Unit) {
             setupVersionCatalog()
         }
         extensions.findByType<MavenPublishBaseExtension>()?.apply {
+            configureBasedOnAppliedPlugins(sourcesJar = true, javadocJar = System.getenv("RUNNING_ON_CI") == "true")
             publishToMavenCentral(SonatypeHost.CENTRAL_PORTAL, automaticRelease = true)
             if (System.getenv("ORG_GRADLE_PROJECT_signingInMemoryKey")?.isNotBlank() == true) {
                 signAllPublications()