+++ reactivestate-android/src/androidMain/kotlin/com/ensody/reactivestate/AutoRunLiveData.kt
@@ -5,12 +5,12 @@ import androidx.lifecycle.Observer
 
 /** Returns [LiveData.getValue] and tracks the observable. */
 public fun <T> Resolver.get(data: LiveData<T>): T? =
-    track(data) { LiveDataObservable(data, autoRunner) }.value
+    track(data) { LiveDataObservable(data) }.value
 
 private class LiveDataObservable<T>(
     private val data: LiveData<T>,
-    autoRunner: BaseAutoRunner,
 ) : AutoRunnerObservable<T?> {
+    private lateinit var autoRunner: BaseAutoRunner
     private var ignore = false
     private val observer = Observer<T> {
         if (!ignore) {
@@ -20,14 +20,15 @@ private class LiveDataObservable<T>(
 
     override val value: T? get() = data.value
 
-    override fun addObserver() {
+    override fun addObserver(autoRunner: BaseAutoRunner) {
         // Prevent recursion and assume the value is already set correctly
         ignore = true
+        this.autoRunner = autoRunner
         data.observeForever(observer)
         ignore = false
     }
 
-    override fun removeObserver() {
+    override fun removeObserver(autoRunner: BaseAutoRunner) {
         data.removeObserver(observer)
     }
 }